+++ packages/core/src/types/index.ts
@@ -98,12 +98,14 @@ export type {
   IImagesLoader,
 } from './IRetrieveConfiguration';
 import type { ImageLoadListener } from './ImageLoadListener';
+import type { Color, ColorLUT } from './Color';
 import type VideoViewportProperties from './VideoViewportProperties';
 import type IVideoViewport from './IVideoViewport';
 import type {
   InternalVideoCamera,
   VideoViewportInput,
 } from './VideoViewportTypes';
+import type BoundsIJK from './BoundsIJK';
 
 export type {
   // config
@@ -211,4 +213,7 @@ export type {
   // video
   InternalVideoCamera,
   VideoViewportInput,
+  BoundsIJK,
+  Color,
+  ColorLUT,
 };