+++ packages/tools/src/tools/annotation/EllipticalROITool.ts
@@ -785,9 +785,7 @@ class EllipticalROITool extends AnnotationTool {
       const rotation = Math.abs(
         viewport.getRotation() - (data.initialRotation || 0)
       );
-      let canvasCorners;
-
-      canvasCorners = <Array<Types.Point2>>(
+      const canvasCorners = <Array<Types.Point2>>(
         getCanvasEllipseCorners(canvasCoordinates) // bottom, top, left, right, keep as is
       );