+++ packages/tools/src/tools/segmentation/CircleScissorsTool.ts
@@ -10,6 +10,7 @@ import {
 } from '../../types';
 
 import { fillInsideCircle } from './strategies/fillCircle';
+import { eraseInsideCircle } from './strategies/eraseCircle';
 import { Events } from '../../enums';
 import { drawCircle as drawCircleSvg } from '../../drawingSvg';
 import {
@@ -69,7 +70,7 @@ class CircleScissorsTool extends BaseTool {
       configuration: {
         strategies: {
           FILL_INSIDE: fillInsideCircle,
-          // ERASE_INSIDE: eraseInsideCircle,
+          ERASE_INSIDE: eraseInsideCircle,
         },
         defaultStrategy: 'FILL_INSIDE',
         activeStrategy: 'FILL_INSIDE',
@@ -275,6 +276,7 @@ class CircleScissorsTool extends BaseTool {
       points: data.handles.points,
       viewPlaneNormal,
       viewUp,
+      strategySpecificConfiguration: {},
     };
 
     this.editData = null;